linux: Make kmeta source path unambiguous
The current implementation of kmeta source path is prone to collisions
in case both 5.10 and common directories contain a subdirectory with
the same name (e.g. 5.10/foo/ and common/foo). In case there is a
collision, only one of the directories is pulled in and that leads to
confusing warnings from SCC. Switch FILESEXTRAPATH and SRC_URI around,
so that the SRC_URI always contains the full path to the kmeta, which
is not ambiguous.

Signed-off-by: Marek Vasut <marex@denx.de>