kas: Switch meta-qt5 and meta-clang to master branch
The master branch of meta-qt5 contains up-to-date Qt5 version compatible
with both OE dunfell 3.1 and kirkstone 4.0, switch the recommended branch
to meta-qt5 master branch to pull in the latest version with latest fixes.

The master branch of meta-clang contains up-to-date version of LLVM clang
with both OE dunfell 3.1 and kirkstone 4.0, switch the recommended branch
to meta-clang master branch to pull in the latest version with latest fixes,
which is especially important when running Chromium which requires up to
date clang.

Signed-off-by: Marek Vasut <marex@denx.de>