u-boot: Set UBOOT_LOCALVERSION
Set UBOOT_LOCALVERSION to a string identifying the OE-built U-Boot
patch level. This should make it possible to identify a bootloader
binary installed on a board even if such binary does not include a
form of /etc/build type of unique image identifier.

The format encodes OE build time MACHINE as well as date and extra
version for the unlikely case of multiple releases during one day.

Signed-off-by: Marek Vasut <marex@denx.de>