machine: Switch DH i.MX6 DHSOM based systems to Linux 6.1.y
The Linux 6.1.y is the current LTS release of the Linux kernel.
Switch DH i.MX6 DHSOM based systems to Linux 6.1.y to use the
current up to date release.

This makes use of DTC_FLAGS to pass '-@' parameter to DT compiler
at kernel build time, instead of a dedicated kernel patch. This
parameter is used to generate __symbols__ node in base DT, which
is necessary for DTO application.

Signed-off-by: Marek Vasut <marex@denx.de>