cypress-firmware: Install board specific CLM blob symlink
Install board specific CLM blob symlink so that the brcmfmac driver
can pick it up immediatelly instead of waiting for the fallback to
start and pick generic CLM blob. The fallback takes about a minute
to start, which makes the WiFi MAC interface appear about a minute
after boot. Improve this.

Signed-off-by: Marek Vasut <marex@denx.de>