linux: Turn wireless support into modules on i.MX6ULL DHCOM
Avoid compiling wireless support into the kernel, build most of it as
kernel modules instead. This is because wireless support depends on
various firmware blobs, like the regulatory database blob, bluetooth
and MAC firmware blob, all of which are only available once the root
filesystem is mounted. With wireless support built as modules, those
modules can only be loaded when the root filesystem is mounted, and
therefore when the firmware blobs stored within it are available too.

Signed-off-by: Marek Vasut <marex@denx.de>