linux: Fix i.MX8MP DHCOM PDK2/PDK3 display DTO PWM polarity inversion
The PWM polarity must be set inverted on all PDK2 and PDK3 panel expansion
boards because of an inverter present on the hardware. This is already set
for i.MX6 DHSOM DTOs, align the i.MX8M Plus DHCOM DTOs to match.

Without the fix, the higher the backlight brightness is set, the lower the
actual panel backlight brightness would be. This fix repairs the inversion.

Signed-off-by: Marek Vasut <marex@denx.de>